[P4ADEV-4609] moves legal facts expiration logic to new FileExpirationService
