+++ openapi/p4pa-send-notification.openapi.yaml
@@ -240,12 +240,6 @@ paths:
           required: true
           schema:
             type: string
-        - name: scheduleDateTime
-          in: query
-          required: true
-          schema:
-            type: string
-            format: date-time
       responses:
         '200':
           description: OK
