+++ src/main/java/it/gov/pagopa/pu/send/exception/ExpirationConfigNotFoundException.java
@@ -1,0 +1,10 @@
+package it.gov.pagopa.pu.send.exception;
+
+import it.gov.pagopa.pu.send.util.ErrorCodeConstants;
+
+public class ExpirationConfigNotFoundException extends BaseBusinessException {
+
+  public ExpirationConfigNotFoundException(String message) {
+    super(ErrorCodeConstants.ERROR_CODE_EXPIRATION_CONFIG_NOT_FOUND, message);
+  }
+}
