+++ src/main/java/it/gov/pagopa/pu/send/service/FileExpirationService.java
@@ -1,0 +1,7 @@
+package it.gov.pagopa.pu.send.service;
+
+import it.gov.pagopa.pu.send.dto.generated.FileExpirationResponseDTO;
+
+public interface FileExpirationService {
+  FileExpirationResponseDTO deleteExpiredLegalFacts(String sendNotificationId, String accessToken);
+}
