+++ src/main/java/it/gov/pagopa/pu/send/service/SendNotificationService.java
@@ -4,9 +4,9 @@ import com.mongodb.client.result.UpdateResult;
 import it.gov.pagopa.pu.send.connector.send.generated.dto.LegalFactCategoryDTO;
 import it.gov.pagopa.pu.send.dto.generated.*;
 import it.gov.pagopa.pu.send.enums.NotificationStatus;
+import it.gov.pagopa.pu.send.model.SendNotificationNoPII;
 
 import java.io.InputStream;
-import java.time.OffsetDateTime;
 import java.util.List;
 
 public interface SendNotificationService {
@@ -20,5 +20,5 @@ public interface SendNotificationService {
   UpdateResult updateNotificationStatus(String notificationRequestId, NotificationStatus newStatus);
   void uploadSendLegalFact(String sendNotificationId, LegalFactCategoryDTO category, String fileName, InputStream inputStream);
   List<LegalFactDTO> getLegalFacts(String sendNotificationId);
-  FileExpirationResponseDTO deleteExpiredLegalFacts(String sendNotificationId, OffsetDateTime scheduleDateTime, String accessToken);
+  SendNotificationNoPII findSendNotification(String sendNotificationId);
 }
