+++ src/main/java/it/gov/pagopa/pu/send/service/SendNotificationServiceImpl.java
@@ -1,8 +1,6 @@
 package it.gov.pagopa.pu.send.service;
 
 import com.mongodb.client.result.UpdateResult;
-import it.gov.pagopa.pu.organization.dto.generated.BrokerConfiguration;
-import it.gov.pagopa.pu.send.connector.organization.service.BrokerConfigurationService;
 import it.gov.pagopa.pu.send.connector.send.generated.dto.LegalFactCategoryDTO;
 import it.gov.pagopa.pu.send.connector.workflow.service.WorkflowService;
 import it.gov.pagopa.pu.send.dto.DocumentDTO;
@@ -32,12 +30,8 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.time.OffsetDateTime;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.function.Supplier;
 
 @Slf4j
 @Service
@@ -50,12 +44,11 @@ public class SendNotificationServiceImpl implements SendNotificationService {
   private final String fileShareBaseUrl;
   private final FileStorerService fileStorerService;
   private final SendNotification2SendNotificationDTOMapper sendNotificationDTOMapper;
-  private final BrokerConfigurationService brokerConfigurationService;
 
   public SendNotificationServiceImpl(@Value("${fileshare-public-base-url}") String fileShareBaseUrl,
                                      SendNotificationPIIRepository sendNotificationPIIRepository,
                                      SendNotificationNoPIIRepository sendNotificationNoPIIRepository, CreateNotificationRequest2SendNotificationMapper mapper, WorkflowService workflowService,
-                                     FileStorerService fileStorerService, SendNotification2SendNotificationDTOMapper sendNotificationDTOMapper, BrokerConfigurationService brokerConfigurationService) {
+                                     FileStorerService fileStorerService, SendNotification2SendNotificationDTOMapper sendNotificationDTOMapper) {
     this.fileShareBaseUrl = fileShareBaseUrl;
     this.sendNotificationPIIRepository = sendNotificationPIIRepository;
     this.sendNotificationNoPIIRepository = sendNotificationNoPIIRepository;
@@ -63,7 +56,6 @@ public class SendNotificationServiceImpl implements SendNotificationService {
     this.workflowService = workflowService;
     this.fileStorerService = fileStorerService;
     this.sendNotificationDTOMapper = sendNotificationDTOMapper;
-    this.brokerConfigurationService = brokerConfigurationService;
   }
 
   @Transactional
@@ -167,7 +159,7 @@ public class SendNotificationServiceImpl implements SendNotificationService {
     return findSendNotification(sendNotificationId).getLegalFacts();
   }
 
-  private SendNotificationNoPII findSendNotification(String sendNotificationId) {
+  public SendNotificationNoPII findSendNotification(String sendNotificationId) {
     return sendNotificationNoPIIRepository.findById(sendNotificationId)
       .orElseThrow(() -> new SendNotificationNotFoundException("Notification not found with id: " + sendNotificationId));
   }
@@ -247,50 +239,6 @@ public class SendNotificationServiceImpl implements SendNotificationService {
       throw new DeleteFileException(String.format(
         "Error while deleting file %s for sendNotificationId %s.", filePath.getFileName(), sendNotificationId
       ));
-    }
-  }
-
-  @Override
-  public FileExpirationResponseDTO deleteExpiredLegalFacts(String sendNotificationId, OffsetDateTime scheduledDateTime, String accessToken) {
-    SendNotificationNoPII sendNotification = findSendNotification(sendNotificationId);
-
-    BrokerConfiguration brokerConfiguration = brokerConfigurationService.getBrokerConfigurationByOrganizationId(sendNotification.getOrganizationId(), accessToken);
-    if(brokerConfiguration == null || brokerConfiguration.getLegalFactsExpirationDays()==null) {
-      log.debug("legalFactsExpirationDays is not configured for the organization having organizationId {}", sendNotification.getOrganizationId());
-      return FileExpirationResponseDTO.builder().build();
-    }
-    Long legalFactsExpirationDays = brokerConfiguration.getLegalFactsExpirationDays();
-
-    OffsetDateTime nextFileExpirationDate = sendNotification.getLegalFacts().stream()
-      .filter(legalFact -> legalFact.getStatus() != FileStatus.EXPIRED && legalFact.getDownloadDate() != null)
-      .map(legalFact -> processFile(
-        sendNotification,
-        legalFact.getFileName(),
-        legalFact.getDownloadDate().plusDays(legalFactsExpirationDays),
-        scheduledDateTime,
-        ()-> sendNotificationNoPIIRepository.updateLegalFactStatus(sendNotificationId, legalFact.getFileName(), FileStatus.EXPIRED)))
-      .filter(Objects::nonNull)
-      .max(Comparator.naturalOrder())
-      .orElse(null);
-
-    return FileExpirationResponseDTO.builder()
-      .nextFileExpirationDate(nextFileExpirationDate)
-      .build();
-  }
-
-  private OffsetDateTime processFile(SendNotificationNoPII notification, String fileName, OffsetDateTime fileExpirationDateTime, OffsetDateTime scheduledDateTime, Supplier<UpdateResult> updateAction) {
-    if (scheduledDateTime.isAfter(fileExpirationDateTime)) {
-      deleteFileAndUpdate(notification.getSendNotificationId(), fileName, notification.getOrganizationId(), updateAction);
-      return null;
-    }
-    return fileExpirationDateTime;
-  }
-
-  private void deleteFileAndUpdate(String sendNotificationId, String fileName, Long organizationId, Supplier<UpdateResult> updateAction) {
-    fileStorerService.deleteFromSharedFolder(organizationId, sendNotificationId, fileName);
-    UpdateResult updateResult = updateAction.get();
-    if (updateResult.getModifiedCount() != 1L) {
-      throw new SendNotificationNotFoundException("Cannot update send notification file having sendNotificationId %s filename %s to status %s".formatted(sendNotificationId, fileName, FileStatus.EXPIRED));
     }
   }
 }
