+++ src/main/java/it/gov/pagopa/pu/send/util/ErrorCodeConstants.java
@@ -7,6 +7,7 @@ public class ErrorCodeConstants {
   public static final String ERROR_CODE_NOTIFICATION_NOT_FOUND = "NOTIFICATION_NOT_FOUND";
   public static final String ERROR_CODE_DEBT_POSITION_NOT_FOUND = "DEBT_POSITION_NOT_FOUND";
   public static final String ERROR_CODE_STREAMS_NOT_FOUND = "STREAMS_NOT_FOUND";
+  public static final String ERROR_CODE_EXPIRATION_CONFIG_NOT_FOUND = "EXPIRATION_CONFIG_NOT_FOUND";
 
   public static final String ERROR_CODE_UPLOAD_ERROR = "UPLOAD_ERROR";
   public static final String ERROR_CODE_DELETE_ERROR = "DELETE_ERROR";
