+++ src/test/java/it/gov/pagopa/pu/send/controller/SendNotificationControllerTest.java
@@ -4,6 +4,7 @@ import com.mongodb.client.result.UpdateResult;
 import it.gov.pagopa.pu.send.connector.send.generated.dto.LegalFactCategoryDTO;
 import it.gov.pagopa.pu.send.dto.generated.*;
 import it.gov.pagopa.pu.send.enums.NotificationStatus;
+import it.gov.pagopa.pu.send.service.FileExpirationService;
 import it.gov.pagopa.pu.send.service.SendNotificationService;
 import it.gov.pagopa.pu.send.util.SecurityUtilsTest;
 import it.gov.pagopa.pu.send.util.TestUtils;
@@ -23,7 +24,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import uk.co.jemos.podam.api.PodamFactory;
 
-import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.stream.Stream;
 
@@ -32,6 +32,8 @@ class SendNotificationControllerTest {
 
   @Mock
   private SendNotificationService sendNotificationServiceMock;
+  @Mock
+  private FileExpirationService fileExpirationServiceMock;
 
   @InjectMocks
   private SendNotificationController sendNotificationController;
@@ -47,6 +49,7 @@ class SendNotificationControllerTest {
   @AfterEach
   void clearContext(){
     SecurityUtilsTest.clearSecurityContext();
+    Mockito.verifyNoMoreInteractions(sendNotificationServiceMock,fileExpirationServiceMock);
   }
 
   @Test
@@ -324,12 +327,11 @@ class SendNotificationControllerTest {
   @Test
   void whenDeleteExpiredLegalFactsThenOk(){
     String sendNotificationId = "SENDNOTIFICATIONID";
-    OffsetDateTime scheduledDateTime = OffsetDateTime.now();
     FileExpirationResponseDTO expectedResponse = podamFactory.manufacturePojo(FileExpirationResponseDTO.class);
 
-    Mockito.when(sendNotificationServiceMock.deleteExpiredLegalFacts(sendNotificationId, scheduledDateTime, accessToken)).thenReturn(expectedResponse);
+    Mockito.when(fileExpirationServiceMock.deleteExpiredLegalFacts(sendNotificationId, accessToken)).thenReturn(expectedResponse);
 
-    ResponseEntity<FileExpirationResponseDTO> response = sendNotificationController.deleteExpiredLegalFacts(sendNotificationId, scheduledDateTime);
+    ResponseEntity<FileExpirationResponseDTO> response = sendNotificationController.deleteExpiredLegalFacts(sendNotificationId);
 
     Assertions.assertEquals(HttpStatus.OK, response.getStatusCode());
     Assertions.assertEquals(expectedResponse, response.getBody());
