+++ src/test/java/it/gov/pagopa/pu/send/service/SendNotificationServiceImplTest.java
@@ -2,8 +2,6 @@ package it.gov.pagopa.pu.send.service;
 
 import com.mongodb.client.result.UpdateResult;
 import it.gov.pagopa.pu.common.pii.citizen.model.PersonalData;
-import it.gov.pagopa.pu.organization.dto.generated.BrokerConfiguration;
-import it.gov.pagopa.pu.send.connector.organization.service.BrokerConfigurationService;
 import it.gov.pagopa.pu.send.connector.send.generated.dto.LegalFactCategoryDTO;
 import it.gov.pagopa.pu.send.connector.workflow.service.WorkflowService;
 import it.gov.pagopa.pu.send.dto.DocumentDTO;
@@ -19,7 +17,7 @@ import it.gov.pagopa.pu.send.mapper.SendNotification2SendNotificationDTOMapper;
 import it.gov.pagopa.pu.send.model.SendNotificationNoPII;
 import it.gov.pagopa.pu.send.repository.SendNotificationNoPIIRepository;
 import it.gov.pagopa.pu.send.repository.SendNotificationPIIRepository;
-import it.gov.pagopa.pu.send.util.TestUtils;
+import it.gov.pagopa.pu.send.util.ErrorCodeConstants;
 import it.gov.pagopa.pu.workflowhub.dto.generated.WorkflowCreatedDTO;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -29,14 +27,12 @@ import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.co.jemos.podam.api.PodamFactory;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -46,8 +42,6 @@ import static org.mockito.Mockito.mockStatic;
 
 @ExtendWith(MockitoExtension.class)
 class SendNotificationServiceImplTest {
-  private final PodamFactory podamFactory = TestUtils.getPodamFactory();
-
   @Mock
   private SendNotificationNoPIIRepository sendNotificationNoPIIRepositoryMock;
   @Mock
@@ -60,8 +54,6 @@ class SendNotificationServiceImplTest {
   private FileStorerService fileStorerServiceMock;
   @Mock
   private SendNotification2SendNotificationDTOMapper sendNotificationDTOMapperMock;
-  @Mock
-  private BrokerConfigurationService brokerConfigurationServiceMock;
 
   @InjectMocks
   private SendNotificationServiceImpl sendNotificationService;
@@ -507,279 +499,26 @@ class SendNotificationServiceImplTest {
   }
 
   @Test
-  void givenNoSendNotificationWhenDeleteExpiredLegalFactsThenThrowSendNotificationNotFoundException() {
-    String accessToken = "accessToken";
-    String sendNotificationId = "sendNotificationId";
-    OffsetDateTime scheduledDateTime = OffsetDateTime.now();
+  void whenFindSendNotificationThenOk() {
+    String sendNotificationId = "123";
+    SendNotificationNoPII expectedResult = new SendNotificationNoPII();
+    expectedResult.setSendNotificationId(sendNotificationId);
 
-    Mockito.when(sendNotificationNoPIIRepositoryMock.findById(sendNotificationId))
-      .thenReturn(Optional.empty());
+    Mockito.when(sendNotificationNoPIIRepositoryMock.findById(sendNotificationId)).thenReturn(Optional.of(expectedResult));
 
-    Assertions.assertThrows(SendNotificationNotFoundException.class, () ->
-      sendNotificationService.deleteExpiredLegalFacts(sendNotificationId, scheduledDateTime, accessToken));
+    SendNotificationNoPII result = sendNotificationService.findSendNotification(sendNotificationId);
+
+    Assertions.assertEquals(expectedResult, result);
   }
 
   @Test
-  void givenNullBrokerConfigurationWhenDeleteExpiredLegalFactsThenReturnNullNextSchedule() {
-    String accessToken = "accessToken";
-    String sendNotificationId = "sendNotificationId";
-    long organizationId = 1L;
-    OffsetDateTime scheduledDateTime = OffsetDateTime.now();
+  void givenNoNotificationWhenFindSendNotificationThen() {
+    String sendNotificationId = "123";
 
-    SendNotificationNoPII notification = new SendNotificationNoPII();
-    notification.setSendNotificationId(sendNotificationId);
-    notification.setOrganizationId(organizationId);
+    Mockito.when(sendNotificationNoPIIRepositoryMock.findById(sendNotificationId)).thenReturn(Optional.empty());
 
-    Mockito.when(sendNotificationNoPIIRepositoryMock.findById(sendNotificationId))
-      .thenReturn(Optional.of(notification));
-    Mockito.when(brokerConfigurationServiceMock.getBrokerConfigurationByOrganizationId(organizationId, accessToken))
-      .thenReturn(null);
+    SendNotificationNotFoundException sendNotificationNotFoundException = Assertions.assertThrows(SendNotificationNotFoundException.class, () -> sendNotificationService.findSendNotification(sendNotificationId));
 
-    FileExpirationResponseDTO result =
-      sendNotificationService.deleteExpiredLegalFacts(sendNotificationId, scheduledDateTime, accessToken);
-
-    Assertions.assertNotNull(result);
-    Assertions.assertNull(result.getNextFileExpirationDate());
-  }
-
-  @Test
-  void givenBrokerConfigurationWithNullExpirationDaysWhenDeleteExpiredLegalFactsThenReturnNullNextSchedule() {
-    String accessToken = "accessToken";
-    String sendNotificationId = "sendNotificationId";
-    long organizationId = 1L;
-    OffsetDateTime scheduledDateTime = OffsetDateTime.now();
-
-    SendNotificationNoPII notification = new SendNotificationNoPII();
-    notification.setSendNotificationId(sendNotificationId);
-    notification.setOrganizationId(organizationId);
-
-    BrokerConfiguration brokerConfiguration = new BrokerConfiguration();
-    brokerConfiguration.setLegalFactsExpirationDays(null);
-
-    Mockito.when(sendNotificationNoPIIRepositoryMock.findById(sendNotificationId))
-      .thenReturn(Optional.of(notification));
-    Mockito.when(brokerConfigurationServiceMock.getBrokerConfigurationByOrganizationId(organizationId, accessToken))
-      .thenReturn(brokerConfiguration);
-
-    FileExpirationResponseDTO result =
-      sendNotificationService.deleteExpiredLegalFacts(sendNotificationId, scheduledDateTime, accessToken);
-
-    Assertions.assertNotNull(result);
-    Assertions.assertNull(result.getNextFileExpirationDate());
-  }
-
-  @Test
-  void givenNoUpdateWhenDeleteExpiredLegalFactsThenThrowSendNotificationNotFoundException() {
-    String accessToken = "accessToken";
-    String sendNotificationId = "sendNotificationId";
-    long expirationDays = 30L;
-    long organizationId = 1L;
-    OffsetDateTime scheduledDateTime = OffsetDateTime.now();
-    OffsetDateTime downloadDate = OffsetDateTime.now().minusDays(60);
-    String fileName = "expired.pdf";
-
-    LegalFactDTO legalFact = LegalFactDTO.builder()
-      .fileName(fileName)
-      .url("some/url")
-      .category(LegalFactCategoryDTO.SENDER_ACK)
-      .status(FileStatus.READY)
-      .downloadDate(downloadDate)
-      .build();
-
-    SendNotificationNoPII notification = new SendNotificationNoPII();
-    notification.setSendNotificationId(sendNotificationId);
-    notification.setOrganizationId(organizationId);
-    notification.setLegalFacts(List.of(legalFact));
-
-    BrokerConfiguration brokerConfiguration = new BrokerConfiguration();
-    brokerConfiguration.setLegalFactsExpirationDays(expirationDays);
-
-    UpdateResult failedUpdateResult = UpdateResult.acknowledged(1, 0L, null);
-
-    Mockito.when(sendNotificationNoPIIRepositoryMock.findById(sendNotificationId))
-      .thenReturn(Optional.of(notification));
-    Mockito.when(brokerConfigurationServiceMock.getBrokerConfigurationByOrganizationId(organizationId, accessToken))
-      .thenReturn(brokerConfiguration);
-    Mockito.when(sendNotificationNoPIIRepositoryMock.updateLegalFactStatus(sendNotificationId, fileName, FileStatus.EXPIRED))
-      .thenReturn(failedUpdateResult);
-
-    Assertions.assertThrows(SendNotificationNotFoundException.class, () ->
-      sendNotificationService.deleteExpiredLegalFacts(sendNotificationId, scheduledDateTime, accessToken));
-  }
-
-  @Test
-  void givenLegalFactAlreadyExpiredWhenDeleteExpiredLegalFactsThenSkipItAndReturnNullNextSchedule() {
-    String accessToken = "accessToken";
-    String sendNotificationId = "sendNotificationId";
-    long expirationDays = 30L;
-    long organizationId = 1L;
-    OffsetDateTime scheduledDateTime = OffsetDateTime.now();
-
-    LegalFactDTO expiredLegalFact = LegalFactDTO.builder()
-      .fileName("already-expired.pdf")
-      .status(FileStatus.EXPIRED)
-      .downloadDate(OffsetDateTime.now().minusDays(60))
-      .build();
-
-    SendNotificationNoPII notification = new SendNotificationNoPII();
-    notification.setSendNotificationId(sendNotificationId);
-    notification.setOrganizationId(organizationId);
-    notification.setLegalFacts(List.of(expiredLegalFact));
-
-    BrokerConfiguration brokerConfiguration = new BrokerConfiguration();
-    brokerConfiguration.setLegalFactsExpirationDays(expirationDays);
-
-    Mockito.when(sendNotificationNoPIIRepositoryMock.findById(sendNotificationId))
-      .thenReturn(Optional.of(notification));
-    Mockito.when(brokerConfigurationServiceMock.getBrokerConfigurationByOrganizationId(organizationId, accessToken))
-      .thenReturn(brokerConfiguration);
-
-    FileExpirationResponseDTO result =
-      sendNotificationService.deleteExpiredLegalFacts(sendNotificationId, scheduledDateTime, accessToken);
-
-    Assertions.assertNotNull(result);
-    Assertions.assertNull(result.getNextFileExpirationDate());
-  }
-
-  @Test
-  void givenLegalFactWithNullDownloadDateWhenDeleteExpiredLegalFactsThenSkipItAndReturnNullNextSchedule() {
-    String accessToken = "accessToken";
-    String sendNotificationId = "sendNotificationId";
-    long expirationDays = 30L;
-    long organizationId = 1L;
-    OffsetDateTime scheduledDateTime = OffsetDateTime.now();
-
-    LegalFactDTO legalFactNoDownload = LegalFactDTO.builder()
-      .fileName("no-download.pdf")
-      .status(FileStatus.READY)
-      .downloadDate(null)
-      .build();
-
-    SendNotificationNoPII notification = new SendNotificationNoPII();
-    notification.setSendNotificationId(sendNotificationId);
-    notification.setOrganizationId(organizationId);
-    notification.setLegalFacts(List.of(legalFactNoDownload));
-
-    BrokerConfiguration brokerConfiguration = new BrokerConfiguration();
-    brokerConfiguration.setLegalFactsExpirationDays(expirationDays);
-
-    Mockito.when(sendNotificationNoPIIRepositoryMock.findById(sendNotificationId))
-      .thenReturn(Optional.of(notification));
-    Mockito.when(brokerConfigurationServiceMock.getBrokerConfigurationByOrganizationId(organizationId, accessToken))
-      .thenReturn(brokerConfiguration);
-
-    FileExpirationResponseDTO result =
-      sendNotificationService.deleteExpiredLegalFacts(sendNotificationId, scheduledDateTime, accessToken);
-
-    Assertions.assertNotNull(result);
-    Assertions.assertNull(result.getNextFileExpirationDate());
-  }
-
-  @Test
-  void givenNoExpiredLegalFactsWhenDeleteExpiredLegalFactsThenReturnNextSchedule() {
-    String accessToken = "accessToken";
-    String sendNotificationId = "sendNotificationId";
-    long expirationDays = 30L;
-    long organizationId = 1L;
-    OffsetDateTime scheduledDateTime = OffsetDateTime.now();
-
-    OffsetDateTime downloadDate1 = OffsetDateTime.now().minusDays(5);
-    OffsetDateTime downloadDate2 = OffsetDateTime.now().minusDays(2);
-    OffsetDateTime downloadDate3 = OffsetDateTime.now().minusDays(10);
-
-    OffsetDateTime expectedNextFileExpirationDate = downloadDate2.plusDays(expirationDays);
-
-    LegalFactDTO legalFact1 = LegalFactDTO.builder()
-      .fileName("file1.pdf")
-      .status(FileStatus.READY)
-      .downloadDate(downloadDate1)
-      .build();
-
-    LegalFactDTO legalFact2 = LegalFactDTO.builder()
-      .fileName("file2.pdf")
-      .status(FileStatus.READY)
-      .downloadDate(downloadDate2)
-      .build();
-
-    LegalFactDTO legalFact3 = LegalFactDTO.builder()
-      .fileName("file3.pdf")
-      .status(FileStatus.READY)
-      .downloadDate(downloadDate3)
-      .build();
-
-    SendNotificationNoPII notification = new SendNotificationNoPII();
-    notification.setSendNotificationId(sendNotificationId);
-    notification.setOrganizationId(organizationId);
-    notification.setLegalFacts(List.of(legalFact1, legalFact2, legalFact3));
-
-    BrokerConfiguration brokerConfiguration = new BrokerConfiguration();
-    brokerConfiguration.setLegalFactsExpirationDays(expirationDays);
-
-    Mockito.when(sendNotificationNoPIIRepositoryMock.findById(sendNotificationId))
-      .thenReturn(Optional.of(notification));
-    Mockito.when(brokerConfigurationServiceMock.getBrokerConfigurationByOrganizationId(organizationId, accessToken))
-      .thenReturn(brokerConfiguration);
-
-    FileExpirationResponseDTO result =
-      sendNotificationService.deleteExpiredLegalFacts(sendNotificationId, scheduledDateTime, accessToken);
-
-    Assertions.assertNotNull(result);
-    Assertions.assertEquals(expectedNextFileExpirationDate, result.getNextFileExpirationDate());
-  }
-
-  @Test
-  void givenPartiallyExpiredLegalFactsWhenDeleteExpiredLegalFactsThenReturnMaxOfRemaining() {
-    String accessToken = "accessToken";
-    String sendNotificationId = "sendNotificationId";
-    long expirationDays = 30L;
-    long organizationId = 1L;
-    OffsetDateTime scheduledDateTime = OffsetDateTime.now();
-
-    OffsetDateTime recentDownloadDate = OffsetDateTime.now().minusDays(5);
-    OffsetDateTime olderDownloadDate  = OffsetDateTime.now().minusDays(10);
-    OffsetDateTime expiredDownloadDate = OffsetDateTime.now().minusDays(60);
-
-    OffsetDateTime expectedNextFileExpirationDate = recentDownloadDate.plusDays(expirationDays);
-
-    LegalFactDTO notExpiredFact1 = LegalFactDTO.builder()
-      .fileName("not-expired-1.pdf")
-      .status(FileStatus.READY)
-      .downloadDate(recentDownloadDate)
-      .build();
-
-    LegalFactDTO notExpiredFact2 = LegalFactDTO.builder()
-      .fileName("not-expired-2.pdf")
-      .status(FileStatus.READY)
-      .downloadDate(olderDownloadDate)
-      .build();
-
-    String expiredFileName = "expired.pdf";
-    LegalFactDTO expiredFact = LegalFactDTO.builder()
-      .fileName(expiredFileName)
-      .status(FileStatus.READY)
-      .downloadDate(expiredDownloadDate)
-      .build();
-
-    SendNotificationNoPII notification = podamFactory.manufacturePojo(SendNotificationNoPII.class);
-    notification.setSendNotificationId(sendNotificationId);
-    notification.setOrganizationId(organizationId);
-    notification.setLegalFacts(List.of(notExpiredFact1, notExpiredFact2, expiredFact));
-
-    BrokerConfiguration brokerConfiguration = podamFactory.manufacturePojo(BrokerConfiguration.class);
-    brokerConfiguration.setLegalFactsExpirationDays(expirationDays);
-    UpdateResult updateResult = UpdateResult.acknowledged(1, organizationId, null);
-
-    Mockito.when(sendNotificationNoPIIRepositoryMock.findById(sendNotificationId))
-      .thenReturn(Optional.of(notification));
-    Mockito.when(brokerConfigurationServiceMock.getBrokerConfigurationByOrganizationId(organizationId, accessToken))
-      .thenReturn(brokerConfiguration);
-    Mockito.when(sendNotificationNoPIIRepositoryMock.updateLegalFactStatus(sendNotificationId, expiredFileName, FileStatus.EXPIRED))
-      .thenReturn(updateResult);
-
-    FileExpirationResponseDTO result =
-      sendNotificationService.deleteExpiredLegalFacts(sendNotificationId, scheduledDateTime, accessToken);
-
-    Assertions.assertNotNull(result);
-    Assertions.assertEquals(expectedNextFileExpirationDate, result.getNextFileExpirationDate());
+    Assertions.assertEquals(ErrorCodeConstants.ERROR_CODE_NOTIFICATION_NOT_FOUND, sendNotificationNotFoundException.getCode());
   }
 }
